refactor: reorganize imports and create getContentsAndSaveCsv function for improved structure
