+++ src/lib/get-group-id.ts
@@ -1,5 +1,5 @@
+import { env } from '@/config';
 import * as fs from 'node:fs';
-import { env } from '../config';
 
 const getGroupId = (): string => {
   const groupDataPath = env.GROUP_ID_PATH;
