+++ src/lib/launch-chrome.ts
@@ -1,6 +1,6 @@
-import { spawn } from 'child_process';
-import os from 'os';
-import path from 'path';
+import { spawn } from 'node:child_process';
+import os from 'node:os';
+import path from 'node:path';
 
 export const launchChrome = (): void => {
   const chromePath = `"C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe"`;
