+++ src/modules/commands/save-cookies.ts
@@ -1,8 +1,8 @@
+import { env } from '@/config';
+import { launchChrome } from '@/lib/launch-chrome';
+import { waitForChrome } from '@/lib/wait-for-chrome';
 import fs from 'node:fs';
 import puppeteer from 'puppeteer';
-import { env } from '../../config';
-import { launchChrome } from '../../lib/launch-chrome';
-import { waitForChrome } from '../../lib/wait-for-chrome';
 
 /**
  * Cookieを保存します。
